{{ cookiecutter.name }}/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["{{ cookiecutter.author }} <{{ cookiecutter.email }}>"]
description = "{{ cookiecutter.description }}"
name = "{{ cookiecutter.name }}"
package-mode = false
readme = "README.md"
version = "0.0.0-dev"

[tool.poetry.dependencies]
bcrypt = "^4.0.1"
django = "^4.2.7"
django-axes = "^6.1.1"
django-behaviors = "^0.5.1"
django-environ = "^0.11.2"
django-filter = "^23.4"
django-healthchecks = "^1.5.0"
django-ipware = "^6.0.0"
django-split-settings = "^1.2.0"
django-storages = "^1.14.2"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.4.2"
drf-jwt = "^1.19.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.5"}
pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
python = "~3.11"
redis = "^5.0.1"
sentry-sdk = "^1.37.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.11.0"
django-stubs = "^4.2.6"
djangorestframework-stubs = "^3.14.4"
dotenv-linter = "^0.4.0"
flake8-absolute-import = "^1.0.0.2"
flake8-bugbear = "^23.9.16"
flake8-cognitive-complexity = "^0.1.0"
flake8-django = "^1.4"
flake8-eradicate = "^1.5.0"
flake8-fixme = "^1.1.1"
flake8-pep3101 = "^2.1.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-pyproject = "^1.2.3"
flake8-variables-names = "^0.0.6"
flake8-walrus = "^1.2.0"
freezegun = "^1.2.2"
ipython = "^8.18.0"
isort = "^5.12.0"
jedi = "^0.19.1"
mixer = {extras = ["django"], version = "^7.2.2"}
mypy = "^1.7.1"
pymarkdownlnt = "^0.9.14"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^4.7.0"
pytest-env = "^1.1.1"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.12.0"
pytest-randomly = "^3.15.0"
pytest-xdist = "^3.5.0"
toml-sort = "^0.23.1"
types-freezegun = "^1.1.10"
types-pillow = "^10.1.0.2"

[tool.black]
line_length = 160

[tool.flake8]
exclude = ["__pycache__", "migrations"]
ignore = [
    "E203",  # whitespace before ':'
    "E265",  # block comment should start with '#'
    "E501",  # line too long ({} > {} characters)
    "F811",  # redefinition of unused name from line {}
    "PT001",  # use @pytest.fixture() over @pytest.fixture
    "SIM102",  # use a single if-statement instead of nested if-statements
    "SIM113",  # use enumerate instead of manually incrementing a counter
]
inline-quotes = '"'

[tool.isort]
include_trailing_comma = true
line_length = 160
multi_line_output = 3
src_paths = ["src", "tests"]
use_parentheses = true

[tool.pymarkdown.plugins.md013]
enabled = false

[tool.pymarkdown.plugins.md045]
enabled = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
addopts = ["--reuse-db"]
env = [
    "AXES_ENABLED = False",
    "CELERY_ALWAYS_EAGER = True",
    "CI = 1",
    "DISABLE_THROTTLING = True",
]
filterwarnings = [
    "ignore:.*'rest_framework_jwt.blacklist' defines default_app_config.*You can remove default_app_config.::django",
    "ignore:distutils Version classes are deprecated. Use packaging.version instead.:DeprecationWarning:pytest_freezegun:17",
]
markers = [
    "freeze_time: freezing time marker (pytest-freezegun does not register it)",
]
python_files = ["test*.py"]
pythonpath = ". src"

[tool.tomlsort]
all = true
in_place = true
sort_first = ["tool.poetry"]
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
